// frontend/FreteEmbarcador-master/src/app/gfe/atualizacoes/cadastros/patioPortaria/visitante/visitante.interface.ts
export interface Visitante {
  id?: number;
  filial?: string;
	visitante?: string;
	nome?: string;
	apelido?: string;
	dataImplant?: string;
	cpf?: string;
	rgVisitante?: string;
	orgaoExped?: string;
	funcionario?: string;
	endereco?: string;
	bairro?: string;
	cep?: string;
	nrCidade?: string;
	fone?: string;
	observacao?: string;
	nmCidade?: string;
	uf?: string;
}